maxr1998/pulse4k: Move Combo code to keymap level
