+++ keyboards/maxr1998/pulse4k/keymaps/default/rules.mk
@@ -1,0 +1,1 @@
+COMBO_ENABLE = yes
