+++ keyboards/maxr1998/pulse4k/keymaps/maxr1998/rules.mk
@@ -1,0 +1,1 @@
+COMBO_ENABLE = yes
