+++ keyboards/maxr1998/pulse4k/rules.mk
@@ -13,7 +13,6 @@ ENCODER_ENABLE = yes       # Rotary encoders
 EXTRAKEY_ENABLE = yes      # Audio control and System control
 CONSOLE_ENABLE = yes       # Console for debug
 COMMAND_ENABLE = no        # Commands for debug and configuration
-COMBO_ENABLE = yes         # Key combo feature
 NKRO_ENABLE = yes           # Enable N-Key Rollover
 BACKLIGHT_ENABLE = no      # Enable keyboard backlight functionality
 AUDIO_ENABLE = no          # Audio output
